imgTo3DModel/app/requirements.txt
# 图片转3D模型系统依赖 - 兼容版本组合
# 基础依赖
numpy>=1.21.0
Pillow>=9.0.0
torch>=1.12.0
torchvision>=0.13.0
gradio>=3.40.0

# TripoSR特定依赖
omegaconf==2.3.0
einops==0.7.0
transformers==4.35.0
trimesh==4.0.5
rembg>=2.0.50
huggingface-hub>=0.16.0
imageio[ffmpeg]
xatlas==0.0.9
moderngl==5.10.0

# 3D处理相关
git+https://github.com/tatsy/torchmcubes.git

# 额外工具
tqdm
opencv-python
matplotlib
